[client] Simplify TrackingStreamingCallback usage to address review feedback
