+++ clients/venice-thin-client/src/main/java/com/linkedin/venice/client/store/streaming/TrackingStreamingCallback.java
@@ -1,46 +1,26 @@
 package com.linkedin.venice.client.store.streaming;
 
-import com.linkedin.venice.client.exceptions.VeniceClientException;
 import com.linkedin.venice.client.stats.ClientStats;
 import java.util.Optional;
 
 
 /**
- * This class exposes one more function: {@link #onDeserializationCompletion(Optional, int, int)}, which
+ * This interface exposes one more function: {@link #onDeserializationCompletion(Optional, int, int)}, which
  * will be used for metric tracking.
  */
-public abstract class TrackingStreamingCallback<K, V> extends StreamingCallback<K, V> {
-  private final StreamingCallback<K, V> inner;
-
-  public TrackingStreamingCallback(StreamingCallback<K, V> inner) {
-    this.inner = inner;
-  }
+public interface TrackingStreamingCallback<K, V> extends StreamingCallback<K, V> {
+  Optional<ClientStats> getStats();
 
   /**
-   * This will invoked when any record deserialization happens.
+   * This will be invoked when any record deserialization happens.
    */
-  public abstract void onRecordDeserialized();
+  void onRecordDeserialized();
 
   /**
    * This will be invoked when Venice Client deserialization is done.
-   * @param veniceException
-   * @param successKeyCnt, this param indicates the total number of existing keys
-   * @param duplicateEntryCnt
+   * @param exception
+   * @param successKeyCount, this param indicates the total number of existing keys
+   * @param duplicateEntryCount
    */
-  public abstract void onDeserializationCompletion(
-      Optional<VeniceClientException> veniceException,
-      int successKeyCnt,
-      int duplicateEntryCnt);
-
-  public abstract ClientStats getStats();
-
-  @Override
-  public void onRecordReceived(K key, V value) {
-    inner.onRecordReceived(key, value);
-  }
-
-  @Override
-  public void onCompletion(Optional<Exception> exception) {
-    inner.onCompletion(exception);
-  }
+  void onDeserializationCompletion(Optional<Exception> exception, int successKeyCount, int duplicateEntryCount);
 }
